[DT-3872] Fold the mapper specs in beside the module they cover
Housekeeping against this repo's own conventions: specs mirror the src path,
so the reverse mapper is tested in v2-common-functions.spec.ts rather than a
file of its own, and a plain wait for an element is a findBy query. The
remaining waitFor calls are for a mock call, an input's value, and a hidden
input -- none of which findBy covers.

Trims the comments added on this branch to the ones saying something the
code cannot.
