+++ src/pages/data_submission/v2/studyDatasetDraft.ts
@@ -13,8 +13,7 @@ export class WrongDraftTypeError extends Error {
 
 /**
  * Loads a draft and maps it into the form's model, refusing anything that is not a study/dataset
- * draft. The read endpoint is generic across draft types, so the type has to be checked before the
- * document is mapped — the id it was loaded by says nothing about what it holds.
+ * draft: the read endpoint is generic, so the id says nothing about what the draft holds.
  * @param draftId The draft UUID
  * @returns Promise resolving to the study the form should edit
  */
