+++ src/types/draft.ts
@@ -1,7 +1,6 @@
 /**
- * Wire types for the generic draft endpoints. A draft's document is whatever its type says it is,
- * so `meta.draftType` is what a caller must check before mapping one — never the id it was loaded
- * by or the route that loaded it.
+ * Wire types for the generic draft endpoints. `meta.draftType` is what a caller must check before
+ * mapping a document — never the id it was loaded by or the route that loaded it.
  */
 
 import { STUDY_DATASET_DRAFT_TYPE } from 'src/types/studyTemplate'
@@ -30,10 +29,7 @@ export interface DraftDetail<TDocument = unknown> {
   meta: DraftMeta
 }
 
-/**
- * Whether a loaded draft is one this workflow may map. A draft of another type belongs to another
- * workflow: it is a recoverable load error rather than something to submit or delete from here.
- */
+/** Whether a loaded draft is one this workflow may map. Another type is a load error, not a submission. */
 export const isStudyDatasetDraft = (
   draft: DraftDetail,
 ): draft is DraftDetail<Record<string, unknown>> =>
